replace dash with a pseudo-element
