+++ client/app/bundles/shared/components/Tag.scss
@@ -5,30 +5,33 @@
   padding: 5px 12px;
   text-indent: 4px;
   display: inline-block;
-  margin-right: 5px;
   margin-bottom: 10px;
-
+  
   font-family: Lato;
   font-size: 10px;
   text-align: center;
   letter-spacing: 1.5px;
   text-transform: uppercase;
-
+  
   border-radius: 5px;
   border: 1px solid $container-color;
-
+  
   mix-blend-mode: normal;
   color: $container-color;
-
+  
   &.dark {
     color: $background-color;
     border: 1px;
     background-color: $container-color;
   }
-
+  
   &.normal {
     color: $background-color;
     border: 1px;
     background-color: $transparent;
   }
+}
+
+span + span {
+  margin-left: 5px;
 }
